event-lead-gen/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
gspread>=5.10.0
google-auth>=2.22.0
